pyproject.toml
[project]
name = "django-otp-actions"
version = "0.1.1"
description = "Django app for managing OTP integration"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

authors = [
    { name = "Kendrick Olori", email = "kendrickolori2@gmail.com" },
]

keywords = [
    "django",
    "otp",
    "two-factor",
    "authentication",
    "security",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "cryptography>=30.0",
    "tzdata",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-django",
    "freezegun",
]

[project.urls]
Homepage = "https://github.com/kendrickolori/django-otp-actions"
Repository = "https://github.com/kendrickolori/django-otp-actions"
Issues = "https://github.com/kendrickolori/django-otp-actions/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/otp_actions"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
